CLI-679: Consolidate internal commands under self
