+++ src/Command/App/NewCommand.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Acquia\Cli\Command;
+namespace Acquia\Cli\Command\App;
 
+use Acquia\Cli\Command\CommandBase;
 use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
 use Symfony\Component\Console\Input\InputArgument;
@@ -15,7 +16,7 @@ use Symfony\Component\Filesystem\Path;
  */
 class NewCommand extends CommandBase {
 
-  protected static $defaultName = 'new';
+  protected static $defaultName = 'app:new';
 
   /**
    * {inheritdoc}.
@@ -23,7 +24,8 @@ class NewCommand extends CommandBase {
   protected function configure(): void {
     $this->setDescription('Create a new Drupal project')
       ->addOption('distribution', NULL, InputOption::VALUE_REQUIRED, 'The Composer package name of the Drupal distribution to download')
-      ->addArgument('directory', InputArgument::OPTIONAL, 'The destination directory');
+      ->addArgument('directory', InputArgument::OPTIONAL, 'The destination directory')
+      ->setAliases(['new']);
   }
 
   /**
