+++ tests/phpunit/src/Misc/EnvDbCredsTest.php
@@ -2,7 +2,7 @@
 
 namespace Acquia\Cli\Tests\Misc;
 
-use Acquia\Cli\Command\ClearCacheCommand;
+use Acquia\Cli\Command\Self\ClearCacheCommand;
 use Acquia\Cli\Tests\CommandTestBase;
 use Acquia\Cli\Tests\TestBase;
 use Symfony\Component\Console\Command\Command;
